Create data_migrations table in initial database migration
When the data_migrate gem was introduced in #3701, a migration should
have been added to create the data_migrations table. This would have
prevented us from relying on the rake task `db:migrate:with_data` to
create the data_migrations table whenever it doesn't exist. Relying on
this questionable mechanic is not good.

Since we rely on the presence of the data_migrations table in the
migration
20170306084558_change_repositories_remote_project_name_to_not_null.rb,
we cannot simply add a migration to create the table. The aforementioned
migration would fail in this case. We're forced to create the
data_migrations table before that faulty migration.